+++ lib/renogen/exceptions/yaml_file_blank.rb
@@ -1,0 +1,21 @@
+module Renogen
+  module Exceptions
+    # This is raised when a yaml file change is found but has not contents
+    class YamlFileBlank < Renogen::Exceptions::Base
+      attr_reader :file_path
+
+      def initialize(file_path)
+        @file_path = file_path
+        super
+      end
+
+      # Friendly error message
+      #
+      # @return [String]
+      def message
+        "Error: File contents blank '#{file_path}'"
+      end
+    end
+  end
+end
+